refactor: SectionService 메서드를 LineService 메서드로 이동
https://github.com/next-step/atdd-subway-map/pull/284#discussion_r795460611
