+++ src/main/java/nextstep/subway/applicaion/dto/LineIncludingStationsResponse.java
@@ -3,7 +3,6 @@ package nextstep.subway.applicaion.dto;
 import nextstep.subway.domain.Line;
 
 import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
