+++ src/main/java/nextstep/subway/ui/SectionController.java
@@ -1,6 +1,6 @@
 package nextstep.subway.ui;
 
-import nextstep.subway.applicaion.SectionService;
+import nextstep.subway.applicaion.LineService;
 import nextstep.subway.applicaion.dto.SectionRequest;
 import nextstep.subway.applicaion.dto.SectionResponse;
 import org.springframework.http.ResponseEntity;
@@ -12,23 +12,23 @@ import java.net.URI;
 @RequestMapping("/lines/{lineId}/sections")
 public class SectionController {
 
-    private SectionService sectionService;
+    private LineService lineService;
 
-    public SectionController(SectionService sectionService) {
-        this.sectionService = sectionService;
+    public SectionController(LineService lineService) {
+        this.lineService = lineService;
     }
 
     @PostMapping
     public ResponseEntity<SectionResponse> createSection(@PathVariable Long lineId,
                                                          @RequestBody SectionRequest sectionRequest) {
-        SectionResponse section = sectionService.saveSection(sectionRequest, lineId);
+        SectionResponse section = lineService.saveSection(sectionRequest, lineId);
         return ResponseEntity.created(URI.create("lines/" + lineId + "/sections/" + section.getId()))
                 .body(section);
     }
 
     @DeleteMapping
     public ResponseEntity<Void> deleteSection(@PathVariable Long lineId, @RequestParam Long stationId) {
-        sectionService.deleteSection(lineId, stationId);
+        lineService.deleteSection(lineId, stationId);
         return ResponseEntity.noContent().build();
     }
 }
